make `fl->fx` truncate
Change `fl->fx` to truncate as it converts, which is typically done
anyway by a machine instruction to convert from floating-point to
integer values. This makes `fl->fx` different from `inexact->exact`
or `fl->exact-integer`, but it brings BC and CS in line.

+++ racket/src/racket/src/longdouble/longdouble.h
@@ -150,7 +150,7 @@ XFORM_NONGCING int long_double_available();
 
 # define double_from_long_double(a) (a)
 # define float_from_long_double(a) (a)
-# define int_from_long_double(a) ((int)(a))
+# define int_from_long_double(a) ((intptr_t)(a))
 # define uintptr_from_long_double(a) ((uintptr_t)(a))
 
 # define long_double_plus(a,b) ((a)+(b))
